requirements.txt
opencv-python
numpy
mss
pytesseract
Pillow
gspread
google-auth